Experimental support for Lollipop
baksmali/smali does not support the dex/oat files from Lollipop,
so we have to use oat2dex to try and deoptimize the odex files.

this method fails quite a few times, but that is not the fault of the script -
oat2jar sometimes fails to parse odex files properly.

because oat2jar requires the bootclasspath to be complete before it can deodex,
we have to process /framework before both /app and /priv-app.